#close 30 Improve Migration

+++ database/migrations/2020_12_27_051529_create_contacts_table.php
@@ -13,14 +13,16 @@ class CreateContactsTable extends Migration
      */
     public function up()
     {
-        Schema::create('contacts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->Integer('user_id')->unsigned();
-            $table->string('major',100);
-            $table->string('school',100);
-            $table->string('phone',10);
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if(!Schema::hasTable('contacts')){
+            Schema::create('contacts', function (Blueprint $table) {
+                $table->increments('id');
+                $table->Integer('user_id')->unsigned();
+                $table->string('major',100);
+                $table->string('school',100);
+                $table->string('phone',10);
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
